aks-desktop: Extract shared EmptyStateCard component
